+++ client/examples/twincat.cpp
@@ -3,15 +3,16 @@
 // Created Date: 19/05/2020
 // Author: Shun Suzuki
 // -----
-// Last Modified: 25/12/2020
+// Last Modified: 01/04/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2020 Hapis Lab. All rights reserved.
 //
 
+#include "link/twincat.hpp"
+
 #include "autd3.hpp"
 #include "runner.hpp"
-#include "link/twincat.hpp"
 
 using namespace std;
 
@@ -19,8 +20,12 @@ int main() {
   auto autd = autd::Controller::Create();
   autd->geometry()->AddDevice(autd::Vector3(0, 0, 0), autd::Vector3(0, 0, 0));
 
-  autd->OpenWith(autd::link::LocalTwinCATLink::Create());
-  if (!autd->is_open()) return ENXIO;
+  try {
+    autd->OpenWith(autd::link::LocalTwinCATLink::Create());
+    if (!autd->is_open()) return ENXIO;
+  } catch (exception& e) {
+    std::cerr << e.what() << std::endl;
+  }
 
   return Run(autd);
 }