+++ client/include/link/debug.hpp
@@ -32,8 +32,8 @@ class DebugLink final : public Link {
   DebugLink(DebugLink&& obj) = delete;
   DebugLink& operator=(DebugLink&& obj) = delete;
 
-  void Open() override;
-  void Close() override;
+  bool Open() override;
+  bool Close() override;
   std::optional<std::string> Send(size_t size, std::unique_ptr<uint8_t[]> buf) override;
   std::optional<std::string> Read(uint8_t* rx, uint32_t buffer_len) override;
   bool is_open() override;