+++ client/include/link/emulator.hpp
@@ -39,8 +39,8 @@ class EmulatorLink final : public Link {
   EmulatorLink(EmulatorLink&& obj) = delete;
   EmulatorLink& operator=(EmulatorLink&& obj) = delete;
 
-  void Open() override;
-  void Close() override;
+  bool Open() override;
+  bool Close() override;
   std::optional<std::string> Send(size_t size, std::unique_ptr<uint8_t[]> buf) override;
   std::optional<std::string> Read(uint8_t* rx, uint32_t buffer_len) override;
   bool is_open() override;